Event requests must be handled before Ping requests do. Otherwise Ping requests may have an impact on the status of an application instance, causing Event request to be ignored and instance update status to be incorrect.
Specific scenario:
An update request is made. An update is served, and processed by the node. Instance reboots, upon restart the instance calls Omaha endpoint with Event Complete 3.2, and a Ping element with the new running CoreOS version.
If the Ping is handled first it will register the instance as running the new CoreOS version, causing the event to be ignored since the version has already been updated. That results in an incorrect state in Coreroller where by the instance is running the new version and has notified Coreroller with event 3.2, but Coreroller did not process the event and did not mark the update as complete.

Test scenario were incorrectly testing this because it wasn't testing this using the Real CoreOS Application ID.